Seperate callbacks into obj and add pause/play effect on hero
