+++ src/styles/07-utils/_fade-up.scss
@@ -10,4 +10,7 @@
     transform: none;
     opacity: 1;
   }
+  [data-animation^="running"] {
+    animation-play-state: running;
+  }
 }
